Show a spinner when a device is being connected or disconnected.
