+++ RGBMasterUWPApp/Pages/DevicesPage.xaml
@@ -4,10 +4,15 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:RGBMasterUWPApp.Pages"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:muxc="using:Microsoft.UI.Xaml.Controls" xmlns:state="using:AppExecutionManager.State" xmlns:common="using:Common"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:muxc="using:Microsoft.UI.Xaml.Controls" xmlns:state="using:AppExecutionManager.State" xmlns:common="using:Common" xmlns:utils="using:RGBMasterUWPApp.Utils"
     mc:Ignorable="d"
     Loaded="Page_Loaded"
    >
+    <Page.Resources>
+        <utils:ConnStatusToProgBarIsActiveConverter x:Key="ConnStatusToProgBarIsActive" />
+        <utils:ConnStatusToProgBarVisibilityConverter x:Key="ConnStatusToProgBarVisibility" />
+        <utils:ConnStatusToCheckboxIsEnabledConverter x:Key="ConnStatusToCheckboxIsEnabled" />
+    </Page.Resources>
     <Grid>
         <VisualStateManager.VisualStateGroups>
             <VisualStateGroup>
@@ -86,7 +91,8 @@
                                     <ListView.ItemTemplate>
                                         <DataTemplate x:DataType="state:DiscoveredDevice">
                                             <StackPanel Orientation="Horizontal">
-                                                <CheckBox x:Name="DeviceSelectionCheckbox" IsChecked="{x:Bind IsChecked, Mode=TwoWay}" Tag="{x:Bind}" Click="CheckBox_Click">
+                                                <ProgressRing Visibility="{x:Bind ConnectionStatus, Mode=OneWay, Converter={StaticResource ConnStatusToProgBarVisibility}}" IsActive="{x:Bind ConnectionStatus, Mode=OneWay, Converter={StaticResource ConnStatusToProgBarIsActive}}" Width="130"/>
+                                                <CheckBox x:Name="DeviceSelectionCheckbox" IsEnabled="{x:Bind ConnectionStatus, Mode=OneWay, Converter={StaticResource ConnStatusToCheckboxIsEnabled}}" IsChecked="{x:Bind IsChecked, Mode=TwoWay}" Tag="{x:Bind}" Click="CheckBox_Click">
                                                     <CheckBox.Resources>
                                                         <muxc:TeachingTip x:Name="TeachingTip_SetName" Title="Set Custom Name" Subtitle="Set a custom name for this device" ActionButtonContent="Save" CloseButtonContent="Cancel" Target="{Binding Change_Device_Name_Button}" IsLightDismissEnabled="True">
                                                             <muxc:TeachingTip.IconSource>
