+++ RGBMasterUWPApp/Pages/EffectsControls/MusicEffectControl.xaml
@@ -8,7 +8,7 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:common="using:Common" xmlns:utils="using:RGBMasterUWPApp.Utils" xmlns:muxc="using:Microsoft.UI.Xaml.Controls" xmlns:state="using:AppExecutionManager.State"
     mc:Ignorable="d">
     <Page.Resources>
-        <utils:DrawingColorToBrushColor x:Key="ColorConverter" />
+        <utils:DrawingColorToBrushColorValueConverter x:Key="ColorConverter" />
     </Page.Resources>
     <StackPanel>
         <ContentDialog x:Name="InfoContentDialog" CloseButtonText="Close">
