+++ RGBMasterUWPApp/RGBMasterUWPApp.csproj
@@ -149,7 +149,10 @@
     <Compile Include="RGBMasterUserControl.xaml.cs">
       <DependentUpon>RGBMasterUserControl.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Utils\DrawingColorToBrushColor.cs" />
+    <Compile Include="Utils\ConnStatusToCheckboxIsEnabledConverter.cs" />
+    <Compile Include="Utils\ConnStatusToProgBarIsActiveConverter.cs" />
+    <Compile Include="Utils\ConnStatusToProgBarVisibilityConverter.cs" />
+    <Compile Include="Utils\DrawingColorToBrushColorValueConverter.cs" />
     <Compile Include="Utils\ExtensionMethods.cs" />
     <EmbeddedResource Include="Properties\RGBMasterUWPApp.rd.xml" />
   </ItemGroup>
