+++ RGBMasterUWPApp/Utils/ConnStatusToProgBarIsActiveConverter.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Common;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -7,18 +8,18 @@ using Windows.UI.Xaml.Data;
 
 namespace RGBMasterUWPApp.Utils
 {
-    public class DrawingColorToBrushColor : IValueConverter
+    public class ConnStatusToProgBarIsActiveConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, string language)
         {
-            var drawingColor = (System.Drawing.Color)value;
-            return Windows.UI.Color.FromArgb(drawingColor.A, drawingColor.R, drawingColor.G, drawingColor.B);
+            var connStatus = (ConnectionStatus)value;
+
+            return connStatus == ConnectionStatus.Connecting || connStatus == ConnectionStatus.Disconnecting;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language)
         {
-            var uiColor = (Windows.UI.Color)value;
-            return System.Drawing.Color.FromArgb(uiColor.A, uiColor.R, uiColor.G, uiColor.B);
+            throw new NotImplementedException();
         }
     }
 }
