+++ RGBMasterUWPApp/Utils/ConnStatusToProgBarVisibilityConverter.cs
@@ -1,0 +1,31 @@
+﻿using Common;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Data;
+
+namespace RGBMasterUWPApp.Utils
+{
+    public class ConnStatusToProgBarVisibilityConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, string language)
+        {
+            var connStatus = (ConnectionStatus)value;
+
+            if (connStatus == ConnectionStatus.Connecting || connStatus == ConnectionStatus.Disconnecting)
+            {
+                return Visibility.Visible;
+            }
+
+            return Visibility.Collapsed;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, string language)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
